chore(developer): rename projectReader to projectRunner
